chore: clean up payments routes, documentation, payments utils
