+++ src/app/routes/api/v3/admin/forms/admin-forms.routes.ts
@@ -8,10 +8,10 @@ import {
 import { AdminFormsFeedbackRouter } from './admin-forms.feedback.routes'
 import { AdminFormsFormRouter } from './admin-forms.form.routes'
 import { AdminFormsLogicRouter } from './admin-forms.logic.routes'
+import { AdminFormsPaymentsRouter } from './admin-forms.payments.routes'
 import { AdminFormsPresignRouter } from './admin-forms.presign.routes'
 import { AdminFormsPreviewRouter } from './admin-forms.preview.routes'
 import { AdminFormsSettingsRouter } from './admin-forms.settings.routes'
-import { AdminFormsPaymentsRouter } from './admin-forms.stripe.routes'
 import { AdminFormsSubmissionsRouter } from './admin-forms.submissions.routes'
 import { AdminFormsTwilioRouter } from './admin-forms.twilio.routes'
